refactor(api): split networks.ts into per-concern sub-routers (A17.5)
The 1196-line api/networks.ts with 19 routes is now api/networks/:
  crud.ts (7)      — list/get/create/patch/leave + sync trigger and history
  members.ts (3)   — add, remove, rotate signing key
  join.ts (4)      — invite, join, join-remote, fork
  topology.ts (3)  — reparent-self, adopt, revert-parent
  votes.ts (2)     — list rounds, cast a vote
  _shared.ts       — only what spans groups: BCRYPT_ROUNDS, SSRF_SAFE_URL,
                     safeMemberList (a schema used by one route stays with it)

index.ts mounts them on the same networksRouter; each declares full paths, so
every URL is unchanged. Also drops one pre-existing dead binding (an unused
rsaPublicKeyPem destructure in join-remote; still schema-validated).

WORTH READING — the scripted extraction silently broke SSRF protection, and the
compiler did not care. A bracket-balance rule truncated the chained declaration

  const SSRF_SAFE_URL = z.string().url().refine(isSsrfSafeUrl, ...)

down to `z.string()`, dropping the guard. It type-checked, built clean, and
happily accepted loopback / IMDS / ULA peer URLs. The red-team suite caught it
(34 failures); nothing else did. Fixed the extractor to end a declaration at its
statement terminator, added a completeness assertion that every non-import source
line survives the move, and left a comment on SSRF_SAFE_URL marking the chain
security-critical. Re-verified red-team 272/272.

Known coupling left as-is: votes.ts imports concludeRoundIfReady from
api/sync.ts — one route module importing domain logic from another. The round
state is config (net.pendingRounds), not in-memory, so the split is safe; giving
that helper a real home belongs to A17.6, which opens sync.ts anyway.

Verified: build clean, no dead code (tsc --noUnusedLocals); red-team 272/272,
sync 191/0, standalone 800/0, brain 183/183, mcp-tools 94/94, proxy-spaces 44/44,
schema 29+85, route-analysis guards 8/8. Docs checked — they document
/api/networks URLs only, which are unchanged, so no doc update needed.
