--- server/src/api/networks.ts
@@ -1,1196 +1,0 @@
-﻿/**
- * Network management API — CRUD for Ythril sync networks.
- *
- * Route prefix: /api/networks
- *
- * Authentication: requireAdmin (PAT Bearer token)
- */
-
-import { Router } from 'express';
-import { v4 as uuidv4 } from 'uuid';
-import bcrypt from 'bcrypt';
-import { z } from 'zod';
-import { requireAdmin } from '../auth/middleware.js';
-import { globalRateLimit } from '../rate-limit/middleware.js';
-import { getConfig, saveConfig, getSecrets, saveSecrets } from '../config/loader.js';
-import { createToken, revokeToken, revokePeerCredentialsIfOrphaned } from '../auth/tokens.js';
-import { createSpace } from '../spaces/spaces.js';
-import { concludeRoundIfReady, sendMemberRemovedNotify } from './sync.js';
-import { getSyncHistory } from '../sync/history.js';
-import { buildBraintreeAncestors } from '../util/braintree.js';
-import { makeSignedOwnCast, forceSetMemberSigningKey } from '../util/signing.js';
-import { log } from '../util/log.js';
-import type { NetworkConfig, NetworkMember, VoteRound } from '../config/types.js';
-
-export const networksRouter = Router();
-
-const BCRYPT_ROUNDS = 12;
-
-// ── SSRF-safe peer URL validation ────────────────────────────────────────────
-// Shared validator from util/ssrf.ts covers:
-//   RFC-1918 IPv4, loopback, 169.254 IMDS, IPv6 ULA (fc00::/7),
-//   IPv6 link-local (fe80::/10), GCP metadata FQDN, embedded credentials.
-import { isSsrfSafeUrl, SSRF_SAFE_MESSAGE } from '../util/ssrf.js';
-
-// ── Schemas ─────────────────────────────────────────────────────────────────
-
-const SSRF_SAFE_URL = z
-  .string()
-  .url()
-  .refine(isSsrfSafeUrl, { message: SSRF_SAFE_MESSAGE });
-
-const CreateNetworkBody = z.object({
-  id: z.string().uuid().optional(),  // optional pre-specified ID for cross-instance registration
-  label: z.string().min(1).max(200),
-  type: z.enum(['closed', 'democratic', 'club', 'braintree', 'pubsub']),
-  spaces: z.array(z.string().min(1)).min(1),
-  votingDeadlineHours: z.number().int().min(1).max(72).default(24),
-  syncSchedule: z.string().optional(),
-  merkle: z.boolean().optional(),
-  requireSignedVotes: z.boolean().optional(),  // strict mode: reject unsigned governance votes
-  myParentInstanceId: z.string().optional(),  // braintree: this instance's parent in the tree (omit → root)
-});
-
-const AddMemberBody = z.object({
-  instanceId: z.string().min(1),
-  label: z.string().min(1).max(200),
-  url: SSRF_SAFE_URL,
-  token: z.string().min(1),   // plaintext peer token — stored as bcrypt hash
-  direction: z.enum(['both', 'push', 'pull']).default('both'),
-  parentInstanceId: z.string().optional(),
-  skipTlsVerify: z.boolean().optional(),
-});
-
-const CastVoteBody = z.object({
-  vote: z.enum(['yes', 'veto']),
-});
-
-const UpdateNetworkBody = z.object({
-  syncSchedule: z.string().optional(),
-  label: z.string().min(1).max(200).optional(),
-  requireSignedVotes: z.boolean().optional(),
-});
-
-const JoinRemoteBody = z.object({
-  /** handshakeId returned by Brain A's POST /api/invite/generate */
-  handshakeId: z.string().uuid(),
-  /** inviteUrl returned by Brain A's POST /api/invite/generate (= Brain A's /api/invite/apply URL) */
-  inviteUrl: z.string().url().refine(isSsrfSafeUrl, { message: SSRF_SAFE_MESSAGE }),
-  /** RSA public key PEM returned by Brain A's POST /api/invite/generate */
-  rsaPublicKeyPem: z.string().min(100),
-  /** Network ID from Brain A's invite bundle */
-  networkId: z.string().uuid(),
-  /** This brain's externally reachable base URL (e.g. https://brain-b.example.com) */
-  myUrl: z.string().url(),
-  /** expiresAt from invite bundle — informational only */
-  expiresAt: z.string().optional(),
-  /** Optional space aliasing: maps remote space IDs to desired local space IDs.
-   *  When the UI detects a collision, the user can choose a different local ID.
-   *  Any remote IDs not present in this map will keep their original ID. */
-  spaceMap: z.record(z.string(), z.string().min(1).max(40).regex(/^[a-z0-9-]+$/)).optional(),
-});
-
-const ReparentSelfBody = z.object({
-  /** instanceId of the new parent (e.g. grandparent) */
-  newParentInstanceId: z.string().uuid(),
-  newParentLabel: z.string().min(1).max(200),
-  newParentUrl: SSRF_SAFE_URL,
-  /** Plaintext token (decrypted from the invite apply response) to call the new parent */
-  tokenForNewParent: z.string().min(1),
-  /** instanceId of the original parent that is offline */
-  originalParentInstanceId: z.string().uuid(),
-});
-
-// ── GET /api/networks ──────────────────────────────────────────────────────
-
-networksRouter.get('/', globalRateLimit, requireAdmin, (_req, res) => {
-  const cfg = getConfig();
-  // Strip sensitive fields
-  const networks = cfg.networks.map(n => ({
-    ...n,
-    members: n.members.map(({ tokenHash: _th, skipTlsVerify: _sv, ...m }) => m),
-    inviteKeyHash: undefined,
-  }));
-  res.json({ networks });
-});
-
-// ── GET /api/networks/:id ──────────────────────────────────────────────────
-
-networksRouter.get('/:id', globalRateLimit, requireAdmin, (req, res) => {
-  const cfg = getConfig();
-  const net = cfg.networks.find(n => n.id === req.params['id']);
-  if (!net) { res.status(404).json({ error: 'Network not found' }); return; }
-
-  const safe = {
-    ...net,
-    members: net.members.map(({ tokenHash: _th, skipTlsVerify: _sv, ...m }) => m),
-    inviteKeyHash: undefined,
-  };
-  res.json(safe);
-});
-
-// ── GET /api/networks/:id/sync-history ─────────────────────────────────────
-
-networksRouter.get('/:id/sync-history', globalRateLimit, requireAdmin, async (req, res) => {
-  try {
-    const cfg = getConfig();
-    const net = cfg.networks.find(n => n.id === req.params['id']);
-    if (!net) { res.status(404).json({ error: 'Network not found' }); return; }
-
-    const limit = Math.min(parseInt(req.query['limit'] as string, 10) || 20, 100);
-    const history = await getSyncHistory(net.id, limit);
-    res.json({ history });
-  } catch (err) {
-    log.error(`GET /api/networks/:id/sync-history: ${err}`);
-    res.status(500).json({ error: 'Internal error' });
-  }
-});
-
-// ── POST /api/networks/:id/sync — manually trigger a sync run ──────────────
-
-networksRouter.post('/:id/sync', globalRateLimit, requireAdmin, (req, res) => {
-  const cfg = getConfig();
-  const net = cfg.networks.find(n => n.id === req.params['id']);
-  if (!net) { res.status(404).json({ error: 'Network not found' }); return; }
-
-  import('../sync/engine.js').then(({ runSyncForNetwork }) => {
-    void runSyncForNetwork(net!.id);
-  }).catch(err => log.error(`POST /api/networks/:id/sync import: ${err}`));
-
-  res.json({ ok: true });
-});
-
-networksRouter.post('/', globalRateLimit, requireAdmin, async (req, res) => {
-  try {
-    const parsed = CreateNetworkBody.safeParse(req.body);
-    if (!parsed.success) { res.status(400).json({ error: parsed.error.message }); return; }
-
-    const { id: presetId, label, type, spaces, votingDeadlineHours, syncSchedule, merkle, requireSignedVotes, myParentInstanceId } = parsed.data;
-    const cfg = getConfig();
-
-    // Validate spaces exist
-    const unknownSpaces = spaces.filter(s => !cfg.spaces.some(cs => cs.id === s));
-    if (unknownSpaces.length > 0) {
-      res.status(400).json({ error: `Unknown spaces: ${unknownSpaces.join(', ')}` });
-      return;
-    }
-
-    // If a preset ID is given, ensure it is not already in use
-    if (presetId && cfg.networks.some(n => n.id === presetId)) {
-      res.status(409).json({ error: 'Network with this ID already exists' });
-      return;
-    }
-
-    const network: NetworkConfig = {
-      id: presetId ?? uuidv4(),
-      label,
-      type,
-      spaces,
-      votingDeadlineHours,
-      syncSchedule,
-      merkle,
-      ...(requireSignedVotes ? { requireSignedVotes: true } : {}),
-      myParentInstanceId: type === 'braintree' ? myParentInstanceId : undefined,
-      members: [],
-      pendingRounds: [],
-      createdAt: new Date().toISOString(),
-    };
-
-    cfg.networks.push(network);
-    saveConfig(cfg);
-
-    log.info(`Created network '${label}' (${type}) id=${network.id}`);
-    const { inviteKeyHash: _ikH, ...safe } = network;
-    res.status(201).json(safe);
-  } catch (err) {
-    log.error(`POST /api/networks: ${err}`);
-    res.status(500).json({ error: 'Internal error' });
-  }
-});
-
-// ── DELETE /api/networks/:id — leave/delete a network ─────────────────────
-
-networksRouter.delete('/:id', globalRateLimit, requireAdmin, async (req, res) => {
-  const cfg = getConfig();
-  const idx = cfg.networks.findIndex(n => n.id === req.params['id']);
-  if (idx < 0) { res.status(404).json({ error: 'Network not found' }); return; }
-
-  const net = cfg.networks[idx]!;
-
-  // Broadcast member_departed to all peers before removing the network locally.
-  const secrets = getSecrets();
-  const warnings: string[] = [];
-  await Promise.all(net.members.map(async (member) => {
-    const peerToken = secrets.peerTokens[member.instanceId];
-    if (!peerToken) return;
-    try {
-      const r = await fetch(`${member.url}/api/notify`, {
-        method: 'POST',
-        headers: {
-          'Authorization': `Bearer ${peerToken}`,
-          'Content-Type': 'application/json',
-        },
-        body: JSON.stringify({ networkId: net.id, instanceId: cfg.instanceId, event: 'member_departed' }),
-        signal: AbortSignal.timeout(5_000),
-      });
-      if (!r.ok) warnings.push(`${member.label}: HTTP ${r.status}`);
-    } catch (err) {
-      const msg = err instanceof Error ? err.message : String(err);
-      log.warn(`member_departed to ${member.label}: ${msg}`);
-      warnings.push(`${member.label}: ${msg}`);
-    }
-  }));
-
-  // Re-fetch config after async peer notifications to avoid clobbering concurrent writes.
-  {
-    const c = getConfig();
-    const i = c.networks.findIndex(n => n.id === req.params['id']);
-    if (i >= 0) { c.networks.splice(i, 1); saveConfig(c); }
-  }
-  log.info(`Deleted network id=${net.id}`);
-
-  // Revoke credentials of peers that no longer share any network with us.
-  for (const member of net.members) {
-    if (member.instanceId === cfg.instanceId) continue;
-    await revokePeerCredentialsIfOrphaned(member.instanceId)
-      .catch(err => log.error(`peer credential revocation for ${member.instanceId}: ${err}`));
-  }
-
-  if (warnings.length) {
-    res.json({ ok: true, warnings });
-  } else {
-    res.status(204).end();
-  }
-});
-
-// ── PATCH /api/networks/:id — update mutable network fields ───────────────
-
-networksRouter.patch('/:id', globalRateLimit, requireAdmin, (req, res) => {
-  const parsed = UpdateNetworkBody.safeParse(req.body);
-  if (!parsed.success) { res.status(400).json({ error: parsed.error.message }); return; }
-
-  const cfg = getConfig();
-  const net = cfg.networks.find(n => n.id === req.params['id']);
-  if (!net) { res.status(404).json({ error: 'Network not found' }); return; }
-
-  if (parsed.data.syncSchedule !== undefined) {
-    net.syncSchedule = parsed.data.syncSchedule || undefined;
-    // Re-register cron timer for this network with the new schedule
-    import('../sync/engine.js').then(({ scheduleSyncForNetwork }) => {
-      scheduleSyncForNetwork(net!.id, net!.syncSchedule);
-    }).catch(err => log.warn(`Failed to reschedule sync for ${net!.id}: ${err}`));
-  }
-  if (parsed.data.label) net.label = parsed.data.label;
-  if (parsed.data.requireSignedVotes !== undefined) net.requireSignedVotes = parsed.data.requireSignedVotes;
-
-  saveConfig(cfg);
-  log.info(`Updated network ${net.id}`);
-  const { inviteKeyHash: _ikh, ...safe } = net;
-  res.json({ ...safe, members: net.members.map(({ tokenHash: _th, skipTlsVerify: _sv, ...m }) => m) });
-});
-
-// ── PUT /api/networks/:id/members/:instanceId/signing-key ──────────────────
-// Break-glass: force-pin a member's governance signing key WITHOUT a rotation
-// proof. Use when a peer lost its old private key (so it cannot produce a
-// continuity proof) and must re-establish trust. Normal rotations propagate
-// automatically via a signed proof over gossip.
-const SigningKeyBody = z.object({ signingPublicKey: z.string().min(100).max(4000) });
-
-networksRouter.put('/:id/members/:instanceId/signing-key', globalRateLimit, requireAdmin, (req, res) => {
-  const parsed = SigningKeyBody.safeParse(req.body);
-  if (!parsed.success) { res.status(400).json({ error: parsed.error.message }); return; }
-  const cfg = getConfig();
-  const net = cfg.networks.find(n => n.id === req.params['id']);
-  if (!net) { res.status(404).json({ error: 'Network not found' }); return; }
-  const member = net.members.find(m => m.instanceId === req.params['instanceId']);
-  if (!member) { res.status(404).json({ error: 'Member not found' }); return; }
-  forceSetMemberSigningKey(member, parsed.data.signingPublicKey);
-  saveConfig(cfg);
-  res.json({ ok: true, instanceId: member.instanceId });
-});
-
-// ── POST /api/networks/join-remote ─────────────────────────────────────────
-// Called by the JOINING brain's UI. Executes the full RSA invite handshake
-// server-side so the browser never handles raw crypto or plaintext tokens.
-//
-// Flow:
-//   1. Brain A admin clicks "Generate invite" → calls POST /api/invite/generate
-//      → gets { handshakeId, inviteUrl, rsaPublicKeyPem, networkId, expiresAt }
-//   2. Brain A admin sends that bundle to Brain B admin (out-of-band)
-//   3. Brain B admin pastes bundle + enters their own URL in Brain B's UI
-//   4. Brain B's UI calls this endpoint
-//   5. This endpoint executes the RSA handshake against Brain A on behalf of Brain B
-//   6. Both sides end up with tokens for each other; network registered locally on Brain B.
-
-networksRouter.post('/join-remote', globalRateLimit, requireAdmin, async (req, res) => {
-  try {
-    const parsed = JoinRemoteBody.safeParse(req.body);
-    if (!parsed.success) { res.status(400).json({ error: parsed.error.message }); return; }
-
-    const { handshakeId, inviteUrl, rsaPublicKeyPem: aPubKeyPem, networkId, myUrl, spaceMap: requestedSpaceMap } = parsed.data;
-    const cfg = getConfig();
-
-    // ── Step A: apply — call Brain A's /api/invite/apply ──────────────────────
-    const { generateKeyPairSync, privateDecrypt, publicEncrypt, constants: C } =
-      await import('node:crypto');
-
-    const { privateKey: bPrivKeyPem, publicKey: bPubKeyPem } = generateKeyPairSync('rsa', {
-      modulusLength: 4096,
-      publicKeyEncoding: { type: 'spki', format: 'pem' },
-      privateKeyEncoding: { type: 'pkcs8', format: 'pem' },
-    });
-
-    let applyRes: Response;
-    try {
-      applyRes = await fetch(inviteUrl, {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify({
-          handshakeId,
-          networkId,
-          instanceId: cfg.instanceId,
-          instanceLabel: cfg.instanceLabel,
-          instanceUrl: myUrl,
-          rsaPublicKeyPem: bPubKeyPem,
-        }),
-        signal: AbortSignal.timeout(15_000),
-      });
-    } catch (err) {
-      log.warn(`join-remote: could not reach ${inviteUrl}: ${err}`);
-      res.status(502).json({ error: `Could not reach inviting brain: ${err}` });
-      return;
-    }
-
-    if (!applyRes.ok) {
-      const errBody = await applyRes.json().catch(() => ({}));
-      res.status(applyRes.status).json(errBody);
-      return;
-    }
-
-    const applyData = await applyRes.json() as {
-      encryptedTokenForB: string;
-      rsaPublicKeyPem: string;
-      instanceId: string;
-      instanceLabel: string;
-      networkId: string;
-      networkLabel: string;
-      networkType: string;
-      spaces: string[];
-    };
-
-    // Decrypt tokenForB — the PAT Brain A created on its own server for Brain B to use
-    let tokenForB: string;
-    try {
-      tokenForB = privateDecrypt(
-        { key: bPrivKeyPem, padding: C.RSA_PKCS1_OAEP_PADDING, oaepHash: 'sha256' },
-        Buffer.from(applyData.encryptedTokenForB, 'base64'),
-      ).toString('utf8');
-    } catch {
-      res.status(400).json({ error: 'Failed to decrypt token from inviting brain' });
-      return;
-    }
-    if (!tokenForB.startsWith('ythril_')) {
-      res.status(400).json({ error: 'Decrypted token has unexpected format' });
-      return;
-    }
-
-    // Create a PAT in this brain's token store scoped to network spaces.
-    // Brain A will present this token when calling THIS brain's sync endpoints.
-    const remoteSpaceIds: string[] = applyData.spaces ?? [];
-    const existingSpaces: string[] = [];
-    const createdSpaces: string[] = [];
-    const spaceMap: Record<string, string> = {};
-
-    for (const remoteId of remoteSpaceIds) {
-      // Check if the user chose a different local ID for this remote space
-      const localId = requestedSpaceMap?.[remoteId] ?? remoteId;
-
-      if (localId !== remoteId) {
-        // Record the alias — sync engine will use this to translate peer space IDs
-        spaceMap[remoteId] = localId;
-      }
-
-      if (cfg.spaces.some(cs => cs.id === localId)) {
-        existingSpaces.push(localId);
-      } else {
-        // Auto-create missing spaces so sync has valid targets.
-        // Label is capitalised version of the slug (e.g. "test" → "Test").
-        try {
-          await createSpace({ id: localId, label: localId.charAt(0).toUpperCase() + localId.slice(1) });
-          createdSpaces.push(localId);
-          log.info(`join-remote: auto-created space '${localId}'${localId !== remoteId ? ` (alias for remote '${remoteId}')` : ''} for network ${networkId}`);
-        } catch (err) {
-          res.status(500).json({ error: `Failed to create space '${localId}': ${err}` });
-          return;
-        }
-      }
-    }
-
-    // All remote spaces now have local counterparts — scope token to local IDs.
-    const allNetworkSpaces = [...existingSpaces, ...createdSpaces];
-    const { record: tokenForARecord, plaintext: tokenForAPlaintext } = await createToken({
-      name: `peer:${applyData.instanceLabel ?? 'remote'}`,
-      expiresAt: null,
-      spaces: allNetworkSpaces.length > 0 ? allNetworkSpaces : undefined,
-      peerInstanceId: applyData.instanceId, // link this PAT to the peer that will present it
-    });
-
-    // ── Step B: finalize — send Brain A an encrypted token for it to call us ──
-    const encryptedTokenForA = publicEncrypt(
-      { key: applyData.rsaPublicKeyPem, padding: C.RSA_PKCS1_OAEP_PADDING, oaepHash: 'sha256' },
-      Buffer.from(tokenForAPlaintext, 'utf8'),
-    ).toString('base64');
-
-    const finalizeUrl = inviteUrl.replace(/\/apply$/, '/finalize');
-    let finalizeRes: Response;
-    try {
-      finalizeRes = await fetch(finalizeUrl, {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify({ handshakeId, encryptedTokenForA }),
-        signal: AbortSignal.timeout(15_000),
-      });
-    } catch (err) {
-      await revokeToken(tokenForARecord.id);
-      res.status(502).json({ error: `Could not finalize with inviting brain: ${err}` });
-      return;
-    }
-
-    if (!finalizeRes.ok) {
-      await revokeToken(tokenForARecord.id);
-      const errBody = await finalizeRes.json().catch(() => ({}));
-      res.status(finalizeRes.status).json(errBody);
-      return;
-    }
-
-    const finalizeData = await finalizeRes.json() as { status: string };
-
-    // ── Register network and peer locally ────────────────────────────────────
-    // Store tokenForB so this brain can call Brain A's sync endpoints.
-    const secrets = getSecrets();
-    secrets.peerTokens[applyData.instanceId] = tokenForB;
-    saveSecrets(secrets);
-
-    // Reload config to get fresh state (apply may have taken a few seconds)
-    const freshCfg = getConfig();
-    let net = freshCfg.networks.find(n => n.id === networkId);
-    if (!net) {
-      net = {
-        id: networkId,
-        label: applyData.networkLabel ?? 'Remote network',
-        type: (applyData.networkType as NetworkConfig['type']) ?? 'closed',
-        spaces: allNetworkSpaces,
-        ...(Object.keys(spaceMap).length > 0 ? { spaceMap } : {}),
-        votingDeadlineHours: 24,
-        members: [],
-        pendingRounds: [],
-        createdAt: new Date().toISOString(),
-        myParentInstanceId: applyData.networkType === 'braintree' ? applyData.instanceId : undefined,
-      };
-      freshCfg.networks.push(net);
-    }
-
-    if (!net.members.some(m => m.instanceId === applyData.instanceId)) {
-      const tokenForAHash = await bcrypt.hash(tokenForAPlaintext, BCRYPT_ROUNDS);
-      net.members.push({
-        instanceId: applyData.instanceId,
-        label: applyData.instanceLabel ?? 'remote',
-        url: new URL(inviteUrl).origin,
-        tokenHash: tokenForAHash,
-        direction: applyData.networkType === 'pubsub' ? 'pull'
-                 : applyData.networkType === 'braintree' ? 'pull'
-                 : 'both',
-        lastSeqReceived: {},
-      });
-    }
-
-    saveConfig(freshCfg);
-    log.info(`join-remote: joined '${applyData.networkLabel}' (${networkId}) via RSA handshake`);
-
-    res.json({
-      status: finalizeData.status ?? 'joined',
-      networkId,
-      networkLabel: applyData.networkLabel,
-      networkType: applyData.networkType,
-      spaces: allNetworkSpaces,
-      existingSpaces,
-      createdSpaces,
-      ...(Object.keys(spaceMap).length > 0 ? { spaceMap } : {}),
-      instanceId: applyData.instanceId,
-      instanceLabel: applyData.instanceLabel,
-    });
-  } catch (err) {
-    log.error(`POST /api/networks/join-remote: ${err}`);
-    res.status(500).json({ error: 'Internal error' });
-  }
-});
-
-// ── POST /api/networks/:id/members — add a peer member ────────────────────
-
-networksRouter.post('/:id/members', globalRateLimit, requireAdmin, async (req, res) => {
-  try {
-    const parsed = AddMemberBody.safeParse(req.body);
-    if (!parsed.success) { res.status(400).json({ error: parsed.error.message }); return; }
-
-    const cfg = getConfig();
-    const net = cfg.networks.find(n => n.id === req.params['id']);
-    if (!net) { res.status(404).json({ error: 'Network not found' }); return; }
-
-    const { instanceId, label, url, token, direction, parentInstanceId, skipTlsVerify } = parsed.data;
-
-    if (net.members.some(m => m.instanceId === instanceId)) {
-      res.status(409).json({ error: 'Member already exists' });
-      return;
-    }
-
-    const tokenHash = await bcrypt.hash(token, BCRYPT_ROUNDS);
-
-    // Re-fetch config after async bcrypt to avoid clobbering concurrent writes.
-    const freshCfg = getConfig();
-    const freshNet = freshCfg.networks.find(n => n.id === req.params['id']);
-    if (!freshNet) { res.status(404).json({ error: 'Network not found' }); return; }
-    if (freshNet.members.some(m => m.instanceId === instanceId)) {
-      res.status(409).json({ error: 'Member already exists' });
-      return;
-    }
-
-    const member: NetworkMember = {
-      instanceId,
-      label,
-      url,
-      tokenHash,
-      direction,
-      parentInstanceId,
-      skipTlsVerify,
-    };
-
-    if (freshNet.type === 'closed' || freshNet.type === 'democratic') {
-      // Open a vote round for the new member
-      const round: VoteRound = {
-        roundId: uuidv4(),
-        type: 'join',
-        subjectInstanceId: instanceId,
-        subjectLabel: label,
-        subjectUrl: url,
-        deadline: new Date(Date.now() + freshNet.votingDeadlineHours * 3_600_000).toISOString(),
-        openedAt: new Date().toISOString(),
-        votes: [],
-        pendingMember: member,
-      };
-      freshNet.pendingRounds.push(round);
-      // Save the plaintext peer token so the sync engine can use it once the vote passes
-      const secrets = getSecrets();
-      secrets.peerTokens[instanceId] = token;
-      saveSecrets(secrets);
-      saveConfig(freshCfg);
-      log.info(`Opened join vote round ${round.roundId} for ${label} in network ${freshNet.id}`);
-      res.status(202).json({ status: 'vote_pending', roundId: round.roundId });
-      return;
-    }
-
-    if (freshNet.type === 'club' || freshNet.type === 'pubsub') {
-      // Club / Pubsub: direct add, no vote required.
-      // Pubsub never allows 'both' — publisher stores subscribers as 'push',
-      // subscriber stores publisher as 'pull'.  If 'both' is provided, default
-      // to 'push' (the common publisher-side case); explicit 'pull' is respected
-      // so the subscriber can manually add the publisher.
-      if (freshNet.type === 'pubsub' && member.direction === 'both') member.direction = 'push';
-      freshNet.members.push(member);
-      const secrets = getSecrets();
-      secrets.peerTokens[instanceId] = token;
-      saveSecrets(secrets);
-      saveConfig(freshCfg);
-      log.info(`Added member ${label} (${instanceId}) to network ${freshNet.id}`);
-      const { tokenHash: _th, skipTlsVerify: _sv, ...safeMember } = member;
-      res.status(201).json(safeMember);
-      return;
-    }
-
-    // Braintree: open a vote round with requiredVoters = ancestry path from self to root.
-    // The proposer (this instance) auto-votes yes.  If the path is only [self] (root case),
-    // concludeRoundIfReady passes immediately and we add the member right away → 201.
-    // Otherwise we return 202 and wait for all ancestors to vote via gossip propagation.
-    const requiredVoters = buildBraintreeAncestors(freshNet, freshCfg.instanceId, freshCfg.instanceId);
-    const round: VoteRound = {
-      roundId: uuidv4(),
-      type: 'join',
-      subjectInstanceId: instanceId,
-      subjectLabel: label,
-      subjectUrl: url,
-      deadline: new Date(Date.now() + freshNet.votingDeadlineHours * 3_600_000).toISOString(),
-      openedAt: new Date().toISOString(),
-      votes: [],
-      pendingMember: member,
-      requiredVoters,
-    };
-    freshNet.pendingRounds.push(round);
-    // Auto-cast this instance's yes vote (proposer implicitly approves their own proposal)
-    round.votes.push(makeSignedOwnCast(freshNet.id, round, freshCfg.instanceId, 'yes'));
-    const secrets = getSecrets();
-    secrets.peerTokens[instanceId] = token;
-    saveSecrets(secrets);
-    const immediatePassed = concludeRoundIfReady(freshNet, round);
-    if (immediatePassed) {
-      // Root case: only self needed to vote → add member directly
-      freshNet.members.push(member);
-      saveConfig(freshCfg);
-      log.info(`Braintree join immediate (root): added ${label} (${instanceId}) to network ${freshNet.id}`);
-      const { tokenHash: _th, skipTlsVerify: _sv, ...safeMember } = member;
-      res.status(201).json(safeMember);
-      return;
-    }
-    saveConfig(freshCfg);
-    log.info(`Opened braintree join round ${round.roundId} for ${label} (${instanceId}) in network ${freshNet.id}`);
-    res.status(202).json({ status: 'vote_pending', roundId: round.roundId });
-  } catch (err) {
-    log.error(`POST /api/networks/:id/members: ${err}`);
-    res.status(500).json({ error: 'Internal error' });
-  }
-});
-
-// ── DELETE /api/networks/:id/members/:instanceId — remove a member ─────────
-
-networksRouter.delete('/:id/members/:instanceId', globalRateLimit, requireAdmin, (req, res) => {
-  const cfg = getConfig();
-  const net = cfg.networks.find(n => n.id === req.params['id']);
-  if (!net) { res.status(404).json({ error: 'Network not found' }); return; }
-
-  const memberIdx = net.members.findIndex(m => m.instanceId === req.params['instanceId']);
-  if (memberIdx < 0) { res.status(404).json({ error: 'Member not found' }); return; }
-
-  if (net.type === 'closed' || net.type === 'democratic') {
-    // Open a remove vote round
-    const member = net.members[memberIdx]!;
-    const round: VoteRound = {
-      roundId: uuidv4(),
-      type: 'remove',
-      subjectInstanceId: member.instanceId,
-      subjectLabel: member.label,
-      subjectUrl: member.url,
-      deadline: new Date(Date.now() + net.votingDeadlineHours * 3_600_000).toISOString(),
-      openedAt: new Date().toISOString(),
-      votes: [],
-    };
-    net.pendingRounds.push(round);
-    saveConfig(cfg);
-    log.info(`Opened remove vote round ${round.roundId} for ${member.label} in network ${net.id}`);
-    res.status(202).json({ status: 'vote_pending', roundId: round.roundId });
-    return;
-  }
-
-  if (net.type === 'club' || net.type === 'pubsub') {
-    // Club / Pubsub: publisher (owner) removes directly, no vote required
-    const removedId = net.members[memberIdx]!.instanceId;
-    net.members.splice(memberIdx, 1);
-    saveConfig(cfg);
-    revokePeerCredentialsIfOrphaned(removedId)
-      .catch(err => log.error(`peer credential revocation for ${removedId}: ${err}`));
-    res.status(204).end();
-    return;
-  }
-
-  // Braintree: open a remove vote round with requiredVoters = ancestor path of the subject.
-  // The subject's parent (and all ancestors up to root) must approve the removal.
-  const subject = net.members[memberIdx]!;
-  // Walk from subject's parent upward; if the subject is a direct child of self,
-  // buildBraintreeAncestors(startId=self) correctly includes self and self's own ancestors.
-  const subjectParentId = subject.parentInstanceId ?? cfg.instanceId;
-  const requiredVoters = buildBraintreeAncestors(net, cfg.instanceId, subjectParentId);
-  const removeRound: VoteRound = {
-    roundId: uuidv4(),
-    type: 'remove',
-    subjectInstanceId: subject.instanceId,
-    subjectLabel: subject.label,
-    subjectUrl: subject.url,
-    deadline: new Date(Date.now() + net.votingDeadlineHours * 3_600_000).toISOString(),
-    openedAt: new Date().toISOString(),
-    votes: [],
-    requiredVoters,
-  };
-  net.pendingRounds.push(removeRound);
-  // Auto-cast this instance's yes vote if we are a required voter
-  if (requiredVoters.includes(cfg.instanceId)) {
-    removeRound.votes.push(makeSignedOwnCast(net.id, removeRound, cfg.instanceId, 'yes'));
-  }
-  const immediatePassed = concludeRoundIfReady(net, removeRound);
-  if (immediatePassed) {
-    // Ancestor path is only [self] → remove immediately (member already spliced by concludeRoundIfReady)
-    saveConfig(cfg);
-    sendMemberRemovedNotify(removeRound.subjectUrl, removeRound.subjectInstanceId, net.id);
-    log.info(`Braintree remove immediate: removed ${subject.label} (${subject.instanceId}) from network ${net.id}`);
-    res.status(204).end();
-    return;
-  }
-  saveConfig(cfg);
-  log.info(`Opened braintree remove round ${removeRound.roundId} for ${subject.label} (${subject.instanceId}) in network ${net.id}`);
-  res.status(202).json({ status: 'vote_pending', roundId: removeRound.roundId });
-});
-
-// ── POST /api/networks/:id/reparent-self ────────────────────────────────────
-// Called by a node on ITSELF after completing the invite apply step.
-// Records the new parent in the local config so this node knows it is
-// temporarily connected to a grandparent rather than its original parent.
-
-networksRouter.post('/:id/reparent-self', globalRateLimit, requireAdmin, async (req, res) => {
-  try {
-    const parsed = ReparentSelfBody.safeParse(req.body);
-    if (!parsed.success) { res.status(400).json({ error: parsed.error.message }); return; }
-
-    const cfg = getConfig();
-    const net = cfg.networks.find(n => n.id === req.params['id']);
-    if (!net) { res.status(404).json({ error: 'Network not found' }); return; }
-    if (net.type !== 'braintree') { res.status(400).json({ error: 'reparent-self is only valid for braintree networks' }); return; }
-
-    const { newParentInstanceId, newParentLabel, newParentUrl, tokenForNewParent, originalParentInstanceId } = parsed.data;
-
-    // Upsert the new parent in the local member list so the engine can report status
-    const existing = net.members.find(m => m.instanceId === newParentInstanceId);
-    if (!existing) {
-      net.members.push({
-        instanceId: newParentInstanceId,
-        label: newParentLabel,
-        url: newParentUrl,
-        tokenHash: '',   // no inbound auth needed on this side — new parent pushes TO us
-        direction: 'push',
-      });
-    }
-
-    // Store the outbound token so this engine can call the new parent if needed
-    const secrets = getSecrets();
-    secrets.peerTokens[newParentInstanceId] = tokenForNewParent;
-    saveSecrets(secrets);
-
-    // Mark the temporary reparent state
-    net.temporaryReparent = {
-      newParentInstanceId,
-      originalParentInstanceId,
-      reparentedAt: new Date().toISOString(),
-    };
-
-    saveConfig(cfg);
-    log.info(`reparent-self: network ${net.id} — new parent ${newParentInstanceId} (was ${originalParentInstanceId})`);
-    res.json({ status: 'reparented', newParentInstanceId, originalParentInstanceId });
-  } catch (err) {
-    log.error(`POST /api/networks reparent-self: ${err}`);
-    res.status(500).json({ error: 'Internal error' });
-  }
-});
-
-// ── POST /api/networks/:id/members/:instanceId/adopt ───────────────────────
-// Called on the GRANDPARENT side. Makes a temporary reparent permanent by clearing
-// originalParentInstanceId from the grandchild's member record.
-
-networksRouter.post('/:id/members/:instanceId/adopt', globalRateLimit, requireAdmin, (req, res) => {
-  const cfg = getConfig();
-  const net = cfg.networks.find(n => n.id === req.params['id']);
-  if (!net) { res.status(404).json({ error: 'Network not found' }); return; }
-
-  const member = net.members.find(m => m.instanceId === req.params['instanceId']);
-  if (!member) { res.status(404).json({ error: 'Member not found' }); return; }
-  if (!member.originalParentInstanceId) {
-    res.status(409).json({ error: 'Member is not in a temporary reparent state' });
-    return;
-  }
-
-  const oldOriginal = member.originalParentInstanceId;
-  delete member.originalParentInstanceId;
-  saveConfig(cfg);
-
-  log.info(`Permanent adoption: '${member.label}' (${member.instanceId}) adopted from ${oldOriginal} in network ${net.id}`);
-  res.json({ status: 'adopted', instanceId: member.instanceId, parentInstanceId: member.parentInstanceId });
-});
-
-// ── POST /api/networks/:id/members/:instanceId/revert-parent ───────────────
-// Called on the GRANDPARENT side when the original parent is back online.
-// Restores the topology: grandchild re-parents to its original parent.
-
-networksRouter.post('/:id/members/:instanceId/revert-parent', globalRateLimit, requireAdmin, (req, res) => {
-  const cfg = getConfig();
-  const net = cfg.networks.find(n => n.id === req.params['id']);
-  if (!net) { res.status(404).json({ error: 'Network not found' }); return; }
-
-  const member = net.members.find(m => m.instanceId === req.params['instanceId']);
-  if (!member) { res.status(404).json({ error: 'Member not found' }); return; }
-  if (!member.originalParentInstanceId) {
-    res.status(409).json({ error: 'Member is not in a temporary reparent state' });
-    return;
-  }
-
-  // Restore original parent
-  const restoredParentId = member.originalParentInstanceId;
-  member.parentInstanceId = restoredParentId;
-  delete member.originalParentInstanceId;
-
-  // Move member from this instance's children back to original parent's children
-  const selfInNet = net.members.find(m => m.instanceId === cfg.instanceId);
-  if (selfInNet?.children) {
-    selfInNet.children = selfInNet.children.filter(c => c !== member.instanceId);
-  }
-  const originalParent = net.members.find(m => m.instanceId === restoredParentId);
-  if (originalParent?.children && !originalParent.children.includes(member.instanceId)) {
-    originalParent.children.push(member.instanceId);
-  }
-
-  // Remove direct outbound token — grandparent no longer pushes directly
-  const secrets = getSecrets();
-  delete secrets.peerTokens[member.instanceId];
-  saveSecrets(secrets);
-
-  saveConfig(cfg);
-  log.info(`Parent reverted: '${member.label}' (${member.instanceId}) re-parented back to ${restoredParentId} in network ${net.id}`);
-  res.json({ status: 'reverted', instanceId: member.instanceId, parentInstanceId: restoredParentId });
-});
-// ── GET /api/networks/:id/vote — list open vote rounds ─────────────────────
-
-networksRouter.get('/:id/votes', globalRateLimit, requireAdmin, (req, res) => {
-  const cfg = getConfig();
-  const net = cfg.networks.find(n => n.id === req.params['id']);
-  if (!net) { res.status(404).json({ error: 'Network not found' }); return; }
-  res.json({ rounds: net.pendingRounds.filter(r => !r.concluded) });
-});
-
-// ── POST /api/networks/:id/votes/:roundId — cast a vote ────────────────────
-
-networksRouter.post('/:id/votes/:roundId', globalRateLimit, requireAdmin, (req, res) => {
-  try {
-    const parsed = CastVoteBody.safeParse(req.body);
-    if (!parsed.success) { res.status(400).json({ error: parsed.error.message }); return; }
-
-    const cfg = getConfig();
-    const net = cfg.networks.find(n => n.id === req.params['id']);
-    if (!net) { res.status(404).json({ error: 'Network not found' }); return; }
-
-    const round = net.pendingRounds.find(r => r.roundId === req.params['roundId'] && !r.concluded);
-    if (!round) { res.status(404).json({ error: 'Round not found or already concluded' }); return; }
-
-    const instanceId = cfg.instanceId;
-    const existing = round.votes.findIndex(v => v.instanceId === instanceId);
-    const cast = makeSignedOwnCast(net.id, round, instanceId, parsed.data.vote);
-    if (existing >= 0) { round.votes[existing] = cast; }
-    else { round.votes.push(cast); }
-
-    concludeRoundIfReady(net, round);
-
-    // If join round concluded and passed, add the pending member —
-    // but only if this instance is the direct parent in the tree (for braintree networks
-    // this check prevents ancestor-voters from adding the member to their own list).
-    if (round.concluded && round.type === 'join' && round.pendingMember &&
-        !net.members.some(m => m.instanceId === round.subjectInstanceId)) {
-      const vetoCount = round.votes.filter(v => v.vote === 'veto').length;
-      const isDirectParent = !round.pendingMember.parentInstanceId ||
-        round.pendingMember.parentInstanceId === cfg.instanceId;
-      if (vetoCount === 0 && (net.type !== 'braintree' || isDirectParent)) {
-        net.members.push(round.pendingMember);
-        log.info(`Join vote ${round.roundId} passed — added member ${round.subjectLabel} to network ${net.id}`);
-      }
-    }
-
-    // If space_deletion round concluded and passed, remove the space on this instance
-    if (round.concluded && round.type === 'space_deletion') {
-      const vetoCount = round.votes.filter(v => v.vote === 'veto').length;
-      if (vetoCount === 0 && round.spaceId) {
-        import('../spaces/spaces.js').then(({ removeSpace }) => {
-          removeSpace(round.spaceId!).catch(err => log.error(`space_deletion vote side-effect: ${err}`));
-        }).catch(err => log.error(`space_deletion import: ${err}`));
-      }
-    }
-
-    // If remove round concluded and passed, notify the ejected member
-    if (round.concluded && round.passed && round.type === 'remove') {
-      sendMemberRemovedNotify(round.subjectUrl, round.subjectInstanceId, net.id);
-    }
-
-    saveConfig(cfg);
-    log.info(`Vote cast in round ${round.roundId}: ${parsed.data.vote} (concluded=${round.concluded})`);
-    res.json({ concluded: round.concluded, round });
-  } catch (err) {
-    log.error(`POST /api/networks votes: ${err}`);
-    res.status(500).json({ error: 'Internal error' });
-  }
-});
-
-// ── POST /api/networks/:id/invite — generate invite key ───────────────────
-
-networksRouter.post('/:id/invite', globalRateLimit, requireAdmin, async (req, res) => {
-  try {
-    const cfg = getConfig();
-    const net = cfg.networks.find(n => n.id === req.params['id']);
-    if (!net) { res.status(404).json({ error: 'Network not found' }); return; }
-
-    const { randomBytes } = await import('crypto');
-    const key = `ythril_invite_${randomBytes(32).toString('base64url')}`;
-    const inviteKeyHash = await bcrypt.hash(key, BCRYPT_ROUNDS);
-    // Re-fetch config after async bcrypt to avoid clobbering concurrent writes.
-    const freshCfg = getConfig();
-    const freshNet = freshCfg.networks.find(n => n.id === req.params['id']);
-    if (!freshNet) { res.status(404).json({ error: 'Network not found' }); return; }
-    freshNet.inviteKeyHash = inviteKeyHash;
-    saveConfig(freshCfg);
-
-    log.info(`Generated new invite key for network ${freshNet.id}${net.type === 'pubsub' ? ' (reusable)' : ' (shown once)'}`);
-    res.json({
-      inviteKey: key,
-      networkId: net.id,
-      ...(net.type === 'pubsub'
-        ? { reusable: true, note: 'This key is reusable — safe to publish in docs, QR codes, or share openly. Regenerating a new key revokes this one.' }
-        : { reusable: false, note: 'Store this key securely — it is single-use and will not be shown again' }),
-    });
-  } catch (err) {
-    log.error(`POST /api/networks/:id/invite: ${err}`);
-    res.status(500).json({ error: 'Internal error' });
-  }
-});
-
-// ── POST /api/networks/:id/join — join via invite key ─────────────────────
-
-const JoinNetworkBody = z.object({
-  inviteKey: z.string().min(1),
-  instanceId: z.string().min(1),
-  label: z.string().min(1).max(200),
-  url: SSRF_SAFE_URL,
-  token: z.string().min(1),  // plaintext token for inbound auth
-  direction: z.enum(['both', 'push', 'pull']).default('both'),
-  parentInstanceId: z.string().optional(),
-  skipTlsVerify: z.boolean().optional(),
-});
-
-/** Member list a joiner receives once admitted (credential fields stripped). */
-function safeMemberList(net: NetworkConfig, excludeInstanceId: string) {
-  return net.members
-    .filter(m => m.instanceId !== excludeInstanceId)
-    .map(({ tokenHash: _th, skipTlsVerify: _sv, ...m }) => m);
-}
-
-networksRouter.post('/:id/join', globalRateLimit, requireAdmin, async (req, res) => {
-  try {
-    const parsed = JoinNetworkBody.safeParse(req.body);
-    if (!parsed.success) { res.status(400).json({ error: parsed.error.message }); return; }
-
-    const cfg = getConfig();
-    const net = cfg.networks.find(n => n.id === req.params['id']);
-    if (!net) { res.status(404).json({ error: 'Network not found' }); return; }
-
-    const keyValid = net.inviteKeyHash
-      ? await bcrypt.compare(parsed.data.inviteKey, net.inviteKeyHash)
-      : false;
-
-    if (!keyValid) {
-      // Vote-governed joins consume the network's invite key when the round opens
-      // and preserve the validated hash on the round record. Re-presenting the
-      // same key lets the joiner poll the outcome of its own round.
-      for (let i = net.pendingRounds.length - 1; i >= 0; i--) {
-        const round = net.pendingRounds[i]!;
-        if (round.type !== 'join' || !round.inviteKeyHash) continue;
-        if (round.subjectInstanceId !== parsed.data.instanceId) continue;
-        if (!await bcrypt.compare(parsed.data.inviteKey, round.inviteKeyHash)) continue;
-
-        if (!round.concluded) {
-          res.status(202).json({ status: 'vote_pending', roundId: round.roundId });
-          return;
-        }
-        if (!round.passed) {
-          res.status(403).json({ error: 'Join was denied by network governance (vetoed or expired)' });
-          return;
-        }
-        // Passed: the member is normally added when the round concludes; re-add
-        // from the round's pendingMember if that side-effect was lost (crash
-        // between conclusion and persistence). Re-fetch config after the async
-        // bcrypt compares to avoid clobbering concurrent writes.
-        const freshCfg = getConfig();
-        const freshNet = freshCfg.networks.find(n => n.id === req.params['id']);
-        if (!freshNet) { res.status(404).json({ error: 'Network not found' }); return; }
-        if (!freshNet.members.some(m => m.instanceId === parsed.data.instanceId)) {
-          if (!round.pendingMember) {
-            res.status(410).json({ error: 'Join round passed but the member record was not retained — generate a new invite' });
-            return;
-          }
-          freshNet.members.push(round.pendingMember);
-          saveConfig(freshCfg);
-        }
-        res.status(200).json({ status: 'joined', members: safeMemberList(freshNet, parsed.data.instanceId), networkId: freshNet.id });
-        return;
-      }
-      if (!net.inviteKeyHash) {
-        res.status(400).json({ error: 'No active invite key — generate one first via POST /invite' });
-        return;
-      }
-      res.status(403).json({ error: 'Invalid invite key' });
-      return;
-    }
-
-    if (net.members.some(m => m.instanceId === parsed.data.instanceId)) {
-      res.status(409).json({ error: 'Member already exists' });
-      return;
-    }
-
-    const { instanceId, label, url, token, direction, parentInstanceId, skipTlsVerify } = parsed.data;
-    const tokenHash = await bcrypt.hash(token, BCRYPT_ROUNDS);
-    // Re-fetch config after async bcrypt to avoid clobbering concurrent writes.
-    const freshCfg = getConfig();
-    const freshNet = freshCfg.networks.find(n => n.id === req.params['id']);
-    if (!freshNet) { res.status(404).json({ error: 'Network not found' }); return; }
-    if (freshNet.members.some(m => m.instanceId === instanceId)) {
-      res.status(409).json({ error: 'Member already exists' });
-      return;
-    }
-    const member: NetworkMember = { instanceId, label, url, tokenHash, direction, parentInstanceId, skipTlsVerify };
-
-    if (freshNet.type === 'closed' || freshNet.type === 'democratic') {
-      const round: VoteRound = {
-        roundId: uuidv4(),
-        type: 'join',
-        subjectInstanceId: instanceId,
-        subjectLabel: label,
-        subjectUrl: url,
-        deadline: new Date(Date.now() + freshNet.votingDeadlineHours * 3_600_000).toISOString(),
-        openedAt: new Date().toISOString(),
-        votes: [],
-        pendingMember: member,             // held here until the vote passes
-        inviteKeyHash: net.inviteKeyHash,  // preserve the original validated hash in the round record
-      };
-      freshNet.pendingRounds.push(round);
-      // Revoke invite key after use to prevent replay
-      freshNet.inviteKeyHash = undefined;
-      // Save the plaintext peer token so the sync engine can use it once the vote passes
-      const secrets = getSecrets();
-      secrets.peerTokens[instanceId] = token;
-      saveSecrets(secrets);
-      saveConfig(freshCfg);
-      log.info(`Join via invite key opened vote round ${round.roundId} for ${label}`);
-      res.status(202).json({ status: 'vote_pending', roundId: round.roundId });
-      return;
-    }
-
-    if (freshNet.type === 'braintree') {
-      // Braintree is vote-governed (S9): the joiner is admitted only after every
-      // ancestor on the path from this (inviting) node to the root votes yes —
-      // same round shape as the admin member-add path. The joiner always becomes
-      // a child of the inviting node; topology fields from the wire are ignored.
-      member.parentInstanceId = freshCfg.instanceId;
-      member.direction = 'push';   // we push to our children
-      const requiredVoters = buildBraintreeAncestors(freshNet, freshCfg.instanceId, freshCfg.instanceId);
-      const round: VoteRound = {
-        roundId: uuidv4(),
-        type: 'join',
-        subjectInstanceId: instanceId,
-        subjectLabel: label,
-        subjectUrl: url,
-        deadline: new Date(Date.now() + freshNet.votingDeadlineHours * 3_600_000).toISOString(),
-        openedAt: new Date().toISOString(),
-        votes: [],
-        pendingMember: member,
-        requiredVoters,
-        inviteKeyHash: net.inviteKeyHash,  // preserve the validated hash so the joiner can poll
-      };
-      freshNet.pendingRounds.push(round);
-      // The inviting node's approval is implicit — it generated the invite key.
-      round.votes.push(makeSignedOwnCast(freshNet.id, round, freshCfg.instanceId, 'yes'));
-      // Consume the key (single-use) and store the peer token for post-admission sync.
-      freshNet.inviteKeyHash = undefined;
-      const secrets = getSecrets();
-      secrets.peerTokens[instanceId] = token;
-      saveSecrets(secrets);
-      const immediatePassed = concludeRoundIfReady(freshNet, round);
-      if (immediatePassed) {
-        // Root case: the ancestor path is only [self] → admit immediately
-        freshNet.members.push(member);
-        saveConfig(freshCfg);
-        log.info(`Braintree join via invite key immediate (root): added ${label} (${instanceId}) to network ${freshNet.id}`);
-        res.status(200).json({ status: 'joined', members: safeMemberList(freshNet, instanceId), networkId: freshNet.id });
-        return;
-      }
-      saveConfig(freshCfg);
-      log.info(`Join via invite key opened braintree ancestor round ${round.roundId} for ${label} (${instanceId}) in network ${freshNet.id}`);
-      res.status(202).json({ status: 'vote_pending', roundId: round.roundId });
-      return;
-    }
-
-    // Club / Pubsub — direct join via invite key (documented behavior)
-    // Pubsub subscribers are always push-only (publisher pushes to them).
-    if (freshNet.type === 'pubsub') member.direction = 'push';
-    freshNet.members.push(member);
-    // Pubsub keys are reusable (publishable in docs, QR codes, etc.)
-    // All other types consume the key after use to prevent replay.
-    if (freshNet.type !== 'pubsub') freshNet.inviteKeyHash = undefined;
-    saveConfig(freshCfg);
-    log.info(`Member ${label} joined network ${freshNet.id} via invite key`);
-
-    // Return peer the member list and network metadata (enough to start syncing)
-    res.status(200).json({ status: 'joined', members: safeMemberList(freshNet, instanceId), networkId: freshNet.id });
-  } catch (err) {
-    log.error(`POST /api/networks/:id/join: ${err}`);
-    res.status(500).json({ error: 'Internal error' });
-  }
-});
-
-// ── POST /api/networks/:id/fork ──────────────────────────────────────────────
-// Creates a new standalone/closed network seeded from the caller's copy of the
-// source network's spaces. Works for:
-//   • Active member  — source network still present; spaces are inherited
-//   • Ejected member — source network is gone (deleted on ejection); caller must
-//     supply spaces explicitly in the request body
-//
-// The source network is never modified. ejectedFromNetworks is never cleared.
-
-const ForkNetworkBody = z.object({
-  label: z.string().min(1).max(200),
-  type: z.enum(['closed', 'club']).default('closed'),
-  votingDeadlineHours: z.number().int().min(1).max(72).optional(),
-  spaces: z.array(z.string().min(1)).optional(),
-});
-
-networksRouter.post('/:id/fork', globalRateLimit, requireAdmin, (req, res) => {
-  try {
-    const parsed = ForkNetworkBody.safeParse(req.body);
-    if (!parsed.success) {
-      res.status(400).json({ error: parsed.error.message });
-      return;
-    }
-
-    const cfg = getConfig();
-    const sourceId = String(req.params['id'] ?? '');
-    const sourceNet = cfg.networks.find(n => n.id === sourceId);
-    const isEjected = cfg.ejectedFromNetworks?.includes(sourceId) ?? false;
-
-    if (!sourceNet && !isEjected) {
-      res.status(404).json({ error: 'Network not found' });
-      return;
-    }
-
-    // Spaces: body override takes precedence; otherwise inherited from source.
-    const spaces = parsed.data.spaces ?? sourceNet?.spaces;
-
-    if (!spaces || spaces.length === 0) {
-      res.status(400).json({
-        error: 'spaces is required when the source network is no longer locally available',
-      });
-      return;
-    }
-
-    // All requested spaces must be locally known.
-    const unknownSpaces = spaces.filter(s => !cfg.spaces.some(cs => cs.id === s));
-    if (unknownSpaces.length > 0) {
-      res.status(400).json({ error: `Unknown spaces: ${unknownSpaces.join(', ')}` });
-      return;
-    }
-
-    const forkedNet: NetworkConfig = {
-      id: uuidv4(),
-      label: parsed.data.label,
-      type: parsed.data.type,
-      spaces,
-      votingDeadlineHours: parsed.data.votingDeadlineHours ?? sourceNet?.votingDeadlineHours ?? 24,
-      members: [],
-      pendingRounds: [],
-      createdAt: new Date().toISOString(),
-    };
-
-    cfg.networks.push(forkedNet);
-    saveConfig(cfg);
-    log.info(`Forked network ${sourceId} → new network ${forkedNet.id} ('${forkedNet.label}')`);
-    res.status(201).json(forkedNet);
-  } catch (err) {
-    log.error(`POST /api/networks/:id/fork: ${err}`);
-    res.status(500).json({ error: 'Internal error' });
-  }
-});
